Actualizador universal

+++ linux_updater/actualizar.sh
@@ -1,51 +1,55 @@
 #! /bin/bash
 
-# Se actualiza la base de datos de los repositorios
-echo "Actualizando listas de repositorios"
-sudo apt-get update
-echo
-
-# Se actualizan los paquetes
-echo "Actualizando paquetes"
-sudo apt-get upgrade -y
-echo
-
-# Se comprueba si hay paquetes retenidos
-echo "Comprobando paquetes retenidos..."
-upgrade_output=$(sudo apt-get upgrade -y)
-if [[ $(echo "$upgrade_output" | grep "retenido") || $(echo "$upgrade_output" | grep "kept back") ]]
+if [[ $(whereis apt-get | grep "/") ]]
 then
-	echo "Se han encontrado paquetes retenidos..."
-	echo "Se procede a su instalación"
-else
-	echo "No se han encontrado paquetes retenidos"
+
+	# Se actualiza la base de datos de los repositorios
+	echo "Actualizando listas de repositorios"
+	sudo apt-get update
+	echo
+
+	# Se actualizan los paquetes
+	echo "Actualizando paquetes"
+	sudo apt-get upgrade -y
+	echo
+
+	# Se comprueba si hay paquetes retenidos
+	echo "Comprobando paquetes retenidos..."
+	upgrade_output=$(sudo apt-get upgrade -y)
+	if [[ $(echo "$upgrade_output" | grep "retenido") || $(echo "$upgrade_output" | grep "kept back") ]]
+	then
+		echo "Se han encontrado paquetes retenidos..."
+		echo "Se procede a su instalación"
+	else
+		echo "No se han encontrado paquetes retenidos"
+	fi
+
+	# Mientras haya paquetes retenidos, se instalan
+	while [[ $(echo "$upgrade_output" | grep "retenido") ]]
+	do
+		sudo apt-get install -y $(echo "$upgrade_output" | awk 'BEGIN { x=0 } ; /retenido/ { x=1 } ; x==2 { print $0 ; x=0 } ; x==1 { x=2 }')
+		upgrade_output=$(sudo apt-get upgrade -y)
+	done
+
+	# Mientras haya paquetes retenidos, se instalan
+	while [[ $(echo "$upgrade_output" | grep "kept back") ]]
+	do
+		sudo apt-get install -y $(echo "$upgrade_output" | awk 'BEGIN { x=0 } ; /kept back/ { x=1 } ; x==2 { print $0 ; x=0 } ; x==1 { x=2 }')
+		upgrade_output=$(sudo apt-get upgrade -y)
+	done
+
+	echo
+
+	# Se desinstalan los paquetes no necesarios
+	echo "Desinstalando paquetes no necesarios..."
+	sudo apt-get autoremove -y --purge
+	echo
+
+	# Se eliminan los paquetes
+	echo "Eliminando paquetes no necesarios..."
+	sudo apt-get autoclean
+	echo
 fi
-
-# Mientras haya paquetes retenidos, se instalan
-while [[ $(echo "$upgrade_output" | grep "retenido") ]]
-do
-	sudo apt-get install -y $(echo "$upgrade_output" | awk 'BEGIN { x=0 } ; /retenido/ { x=1 } ; x==2 { print $0 ; x=0 } ; x==1 { x=2 }')
-	upgrade_output=$(sudo apt-get upgrade -y)
-done
-
-# Mientras haya paquetes retenidos, se instalan
-while [[ $(echo "$upgrade_output" | grep "kept back") ]]
-do
-	sudo apt-get install -y $(echo "$upgrade_output" | awk 'BEGIN { x=0 } ; /kept back/ { x=1 } ; x==2 { print $0 ; x=0 } ; x==1 { x=2 }')
-	upgrade_output=$(sudo apt-get upgrade -y)
-done
-
-echo
-
-# Se desinstalan los paquetes no necesarios
-echo "Desinstalando paquetes no necesarios..."
-sudo apt-get autoremove -y --purge
-echo
-
-# Se eliminan los paquetes
-echo "Eliminando paquetes no necesarios..."
-sudo apt-get autoclean
-echo
 
 # Se actualizan los paquetes de Snap
 if [[ $(whereis snap | grep "/") ]]
